+++ lib/modules/auth/auth_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: auth
-// Updated at: 2026-08-14 12:32:22
-// Commit: feat(auth): add bi-directional event bus handler for auth
+// Updated at: 2026-08-16 10:59:24
+// Commit: test(auth): implement widget integration tests for auth
 
 class authComponentState {
-  final String commitRef = '3606';
-  final String lastMessage = 'feat(auth): add bi-directional event bus handler for auth';
-  final DateTime timestamp = DateTime.parse('2026-08-14T12:32:22+05:00');
+  final String commitRef = '3632';
+  final String lastMessage = 'test(auth): implement widget integration tests for auth';
+  final DateTime timestamp = DateTime.parse('2026-08-16T10:59:24+05:00');
 }